backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
numpy==1.24.3
librosa==0.10.0.post2
transformers==4.29.2
networkx==3.1
pydantic==1.10.8
python-dotenv==1.0.0
pytest==7.3.1
soundfile==0.12.1
matplotlib==3.7.1
langchain==0.0.285
langchain-community==0.0.13
openai==1.3.0  # For OpenAI API integration

# Add dependencies for various LLM providers
requests==2.31.0  # Used by multiple providers for HTTP requests
anthropic==0.5.0  # For Anthropic Claude API integration
google-generativeai==0.3.1  # For Google Gemini API integration
huggingface-hub==0.19.0  # For HuggingFace API integration 